Move naming-conventions tests to separate file
